add status indicator badge and refactor the code
