+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
-    "@orpc/experimental-ratelimit": "^1.14.0"
+    "@orpc/experimental-ratelimit": "^1.14.0",
+    "react-calendar-heatmap": "^1.10.0"
   }
 }
